Modified csio_init to take a parameter to prevent writing of the csio pid file.  Every place calling csio_init in csio has the flag set to true.  Only CresStreamSvc sets the flag to false.
CresStreamSvc was calling it and this was resulting in SIGUSR2 being sent to CresStreamSvc.  Since SIGUSR2 was not being handled, it would terminate CresStreamSvc.  SIGUSR2 is used to indicate that DHCP settings have been changed so that we can send joins back to the control system.

Code Review: Sam, Rob S and John F post commit

git-svn-id: svn://linux2/subversion/Crestron/Android_tst600/trunk/customFiles/src/external/crestron/CresStreamSvc@196407 634c09e3-b06c-4293-a491-0870fed93cb0